Move error routines to errors.c
